+++ CHANGELOG.md
@@ -9,7 +9,10 @@ FlatLaf Change Log
   For example: `mySlider.putClientProperty( "FlatLaf.style", "trackWidth: 2" );`
 - TextField, FormattedTextField and PasswordField: Support leading and trailing
   icons (set client property `JTextField.leadingIcon` or
-  `JTextField.trailingIcon` to an `Icon`). (PR #378; issue #368)
+  `JTextField.trailingIcon` to a `javax.swing.Icon`). (PR #378; issue #368)
+- TextField, FormattedTextField and PasswordField: Support leading and trailing
+  components (set client property `JTextField.leadingComponent` or
+  `JTextField.trailingComponent` to a `java.awt.Component`). (PR #379)
 
 #### Fixed bugs